+++ DevTyr.Gullap/Model/SiteConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace DevTyr.Gullap.Model
+{
+    public class SiteConfiguration
+    {
+        public string Title { get; set; }
+    }
+}