+++ Solution Items/RELEASENOTES.md
@@ -1,5 +1,13 @@
 # RELEASE NOTES
 
+**1.1.0**
+
+- Replaced current file header with a Yaml Front Matter (breaking!!)
+- Removed MenuBuilder (this can be done via templating now)
+- Available variables for templating changed, please see documentation
+- If there is no command argument given for the site path, the current path is used
+- Optional site configuration (Yaml)
+
 **1.0.1**
 
 - Added support of Date